Progressing with programs init
